+++ web/tttool_web/__init__.py
@@ -1,3 +1,3 @@
 """A small web GUI around tttool, the Tiptoi® swiss army knife."""
 
-__version__ = "1.2.0"
+__version__ = "1.3.0"